@@ -1,212 +1,205 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    
-	<parent>
-		<groupId>org.sonatype.oss</groupId>
-		<artifactId>oss-parent</artifactId>
-		<version>7</version>
-	</parent>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.gamepad4j</groupId>
-	<artifactId>gamepad4j-root</artifactId>
-	<version>0.3-SNAPSHOT</version>
-	<name>Java Multiplatform Gamepad API</name>
-	<packaging>pom</packaging>
-	<description>API for using gamepads (Xbox 360, PS3 etc.) with Java.</description>
+  <groupId>org.gamepad4j</groupId>
+  <artifactId>gamepad4j-root</artifactId>
+  <version>0.3-SNAPSHOT</version>
 
-	<licenses>
-		<license>
-			<name>Apache License V2.0</name>
-			<url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
-		</license>
-	</licenses>
+  <packaging>pom</packaging>
+  <modelVersion>4.0.0</modelVersion>
+  <name>Java Multiplatform Gamepad API</name>
+  <description>API for using gamepads (Xbox 360, PS3 etc.) with Java.</description>
+  <licenses>
+    <license>
+      <name>Apache License V2.0</name>
+      <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
+    </license>
+  </licenses>
+  <url>https://github.com/marcelschoen/gamepad4j</url>
+  <scm>
+    <connection>scm:git:git@github.com:marcelschoen/gamepad4j.git</connection>
+    <developerConnection>
+      scm:git:[fetch=]git@github.com:marcelschoen/gamepad4j.git[push=]git@github.com:marcelschoen/gamepad4j.git
+    </developerConnection>
+    <url>git@github.com:marcelschoen/gamepad4j.git</url>
+  </scm>
+  <developers>
+    <developer>
+      <id>msc</id>
+      <name>Marcel Schoen</name>
+      <email>marcelschoen@users.sourceforge.net</email>
+    </developer>
+  </developers>
 
-	<url>https://github.com/marcelschoen/gamepad4j</url>
+  <modules>
+    <!-- PLANNED FOR FUTURE RELEASES
+  <module>gamepad4j-android-2</module>
+  <module>gamepad4j-xperia-play</module>
+  <module>gamepad4j-gamestick</module>
+  <module>gamepad4j-libgdx</module>
+  <module>gamepad4j-ouya</module>
+     -->
+    <module>gamepad4j-api</module>
+    <module>gamepad4j-desktop</module>
+    <module>gamepad4j-uberjar</module>
+  </modules>
 
-	<scm>
-		<connection>scm:git:git@github.com:marcelschoen/gamepad4j.git</connection>
-		<developerConnection>scm:git:[fetch=]git@github.com:marcelschoen/gamepad4j.git[push=]git@github.com:marcelschoen/gamepad4j.git</developerConnection>
-		<url>git@github.com:marcelschoen/gamepad4j.git</url>
-	</scm>
-	
-	<developers>
-		<developer>
-			<id>msc</id>
-			<name>Marcel Schoen</name>
-			<email>marcelschoen@users.sourceforge.net</email>
-		</developer>
-	</developers>
+  <profiles>
+    <profile>
+      <id>release-sign-artifacts</id>
+      <activation>
+        <property>
+          <name>performRelease</name>
+          <value>true</value>
+        </property>
+      </activation>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-gpg-plugin</artifactId>
+            <executions>
+              <execution>
+                <id>sign-artifacts</id>
+                <phase>verify</phase>
+                <goals>
+                  <goal>sign</goal>
+                </goals>
+              </execution>
+            </executions>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
 
-	<modules>
-	    <!-- PLANNED FOR FUTURE RELEASES
-		<module>gamepad4j-android-2</module>
-		<module>gamepad4j-xperia-play</module>
-		<module>gamepad4j-gamestick</module>
-		<module>gamepad4j-libgdx</module>
-	     -->
-		<module>gamepad4j-api</module>
-		<module>gamepad4j-ouya</module>
-		<module>gamepad4j-desktop</module>
-		<module>gamepad4j-uberjar</module>
-	</modules>
-	
-	<profiles>
-		<profile>
-			<id>release-sign-artifacts</id>
-			<activation>
-				<property>
-					<name>performRelease</name>
-					<value>true</value>
-				</property>
-			</activation>
-			<build>
-				<plugins>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-gpg-plugin</artifactId>
-						<version>1.1</version>
-						<executions>
-							<execution>
-								<id>sign-artifacts</id>
-								<phase>verify</phase>
-								<goals>
-									<goal>sign</goal>
-								</goals>
-							</execution>
-						</executions>
-					</plugin>
-				</plugins>
-			</build>
-		</profile>
+    <profile>
+      <id>java-8-gotchas</id>
+      <activation>
+        <jdk>1.8</jdk>
+      </activation>
+      <build>
+        <pluginManagement>
+          <plugins>
+            <plugin>
+              <groupId>org.apache.maven.plugins</groupId>
+              <artifactId>maven-javadoc-plugin</artifactId>
+              <configuration>
+                <additionalJOptions>
+                  <additionalJOption>-J-Duser.language=en_US</additionalJOption>  <!-- TODO ad-hoc -->
+                </additionalJOptions>
+                <additionalOptions>
+                  <additionalOption>-Xdoclint:none</additionalOption>
+                  <additionalOption>-Xdoclint:-missing</additionalOption>
+                </additionalOptions>
+                <locale>en_US</locale> <!-- TODO doesn't work why ??? -->
+                <debug>true</debug>
+                <verbose>false</verbose>
+                <failOnWarnings>false</failOnWarnings>
+                <failOnError>false</failOnError>
+                <quiet>true</quiet>
+              </configuration>
+            </plugin>
+          </plugins>
+        </pluginManagement>
+      </build>
+    </profile>
 
-		<profile>
-			<id>java-8-gotchas</id>
-			<activation>
-				<jdk>1.8</jdk>
-			</activation>
-			<build>
-				<pluginManagement>
-					<plugins>
-						<plugin>
-							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-javadoc-plugin</artifactId>
-							<configuration>
-								<additionalparam>-Xdoclint:none</additionalparam>
-							</configuration>
-						</plugin>
-					</plugins>
-				</pluginManagement>
-			</build>
-		</profile>
+  </profiles>
 
-	</profiles>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-enforcer-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>enforce-maven</id>
+            <goals>
+              <goal>enforce</goal>
+            </goals>
+            <configuration>
+              <rules>
+                <requireMavenVersion>
+                  <version>3.0.5</version>
+                </requireMavenVersion>
+              </rules>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
 
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-enforcer-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>enforce-maven</id>
-						<goals>
-							<goal>enforce</goal>
-						</goals>
-						<configuration>
-							<rules>
-								<requireMavenVersion>
-									<version>3.0.5</version>
-								</requireMavenVersion>
-							</rules>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-<!--
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-gpg-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>sign-artifacts</id>
-						<phase>verify</phase>
-						<goals>
-							<goal>sign</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>		
--->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
-				</configuration>
-			</plugin>
-		</plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+          <encoding>UTF-8</encoding>
+        </configuration>
+      </plugin>
+    </plugins>
 
-		<pluginManagement>
-			<plugins>
-			    <plugin>
-					<groupId>org.codehaus.mojo</groupId>
-					<artifactId>properties-maven-plugin</artifactId>
-					<version>1.0-alpha-1</version>
-			    </plugin>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-enforcer-plugin</artifactId>
-					<version>1.3.1</version>
-				</plugin>
-			    <plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-jar-plugin</artifactId>
-					<version>2.5</version>
-			    </plugin>
-			    <plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-assembly-plugin</artifactId>
-					<version>2.4.1</version>
-			    </plugin>
-			    <plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.1</version>
-			    </plugin>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-gpg-plugin</artifactId>
-					<version>1.5</version>
-				</plugin>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-source-plugin</artifactId>
-					<version>2.3</version>
-				</plugin>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-shade-plugin</artifactId>
-					<version>2.3</version>
-				</plugin>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-javadoc-plugin</artifactId>
-					<version>2.9.1</version>
-				</plugin>
-			</plugins>
-		</pluginManagement>
-		
-	</build>
-		
-	<dependencyManagement>
-	    <dependencies>
-		    <dependency>
-		        <artifactId>junit</artifactId>
-		        <groupId>junit</groupId>
-		        <version>4.10</version>
-		    </dependency>
-	    </dependencies>
-	</dependencyManagement>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.codehaus.mojo</groupId>
+          <artifactId>properties-maven-plugin</artifactId>
+          <version>1.1.0</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-enforcer-plugin</artifactId>
+          <version>3.3.0</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-jar-plugin</artifactId>
+          <version>2.5</version> <!-- use old version because of native plugin -->
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-assembly-plugin</artifactId>
+          <version>3.6.0</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-compiler-plugin</artifactId>
+          <version>3.11.0</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-gpg-plugin</artifactId>
+          <version>3.1.0</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-source-plugin</artifactId>
+          <version>3.2.1</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-shade-plugin</artifactId>
+          <version>3.2.4</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-javadoc-plugin</artifactId>
+          <version>3.5.0</version>
+        </plugin>
+      </plugins>
+    </pluginManagement>
 
+  </build>
+
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.junit</groupId>
+        <artifactId>junit-bom</artifactId>
+        <version>5.10.0</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
 </project>
